Support writing very large arrays, the code will still fail with EXTREMELY large arrays (With length in the order 2**64) but it's a non-issue.

+++ bitstream_python/bitstream.pxd
@@ -19,8 +19,8 @@ cdef class BitStream:
     cdef int64 offsetRead
     cdef int64 offsetWrite
 
-    cpdef write(BitStream self, value, type key, int count=?)
-    cpdef read(BitStream self, type key, int count=?)
+    cpdef write(BitStream self, value, type key, int64 count=?)
+    cpdef read(BitStream self, type key, int64 count=?)
     cpdef readInt(BitStream self, int size, int count=?)
 
     cpdef void seek(BitStream self, int64 offset)
